refactor: categories page decomposition
