+++ src/components/categories/AddCategoryModal.tsx
@@ -1,0 +1,29 @@
+import Modal from 'src/components/modal/Modal.tsx';
+
+interface AddCategoryModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+const AddCategoryModal = ({ isOpen, onClose }: AddCategoryModalProps) => {
+  return (
+    <Modal isOpen={isOpen} onClose={onClose} title='Add New Category'>
+      <div>
+        <p>Add new category form will go here</p>
+        <p>
+          Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam,
+          quos. Lorem ipsum dolor sit amet consectetur adipisicing elit.
+          Quisquam, quos. Lorem ipsum dolor sit amet consectetur adipisicing
+          elit. Quisquam, quos. Lorem ipsum dolor sit amet consectetur
+          adipisicing elit. Quisquam, quos. Lorem ipsum dolor sit amet
+          consectetur adipisicing elit. Quisquam, quos. Lorem ipsum dolor sit
+          amet consectetur adipisicing elit. Quisquam, quos. Lorem ipsum dolor
+          sit amet consectetur adipisicing elit. Quisquam, quos. Lorem ipsum
+          dolor sit amet consectetur adipisicing elit. Quisquam, quos.
+        </p>
+      </div>
+    </Modal>
+  );
+};
+
+export default AddCategoryModal;
